+++ src/CommunityReportFormat.h
@@ -19,6 +19,9 @@ struct CommunityReportFacts
     bool snapshotHealthy = false;
     bool startupCompatible = false;
     bool coreMemoryOptimized = false;
+    bool telemetryAvailable = false;
+    bool telemetryPreferenceKnown = false;
+    bool telemetryEnabled = false;
 };
 
 [[nodiscard]] QString formatCommunityReport(const CommunityReportFacts &facts);